Refactor the Library component to cleanup the App component
